Separate utility functions of keras_model into its own utils.py
